+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+* added support for Click-to-sms and send-sms-by-interrupt
 
 ## [4.0.1]
 * fixed token refresh for SQS